feat(runtime): add process sandbox executor abstraction
